backend/requirements.txt
django-cors-headers==4.7.0
whitenoise
asgiref==3.9.1
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.4.3
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0
dj-database-url==2.1.0
